[refactor] : 소켓 분리 및 수정
